add sprite flip, fix animation, add screen resize
Sprites now have a flipped version that is stored in f_images. This may
become merged with the regular version or a new "Image" wrapper may be
created to store both. A sprite may use the flipped version by enable
the sprite.flip flag.

Animations and physics ticks now properly update their last frame/tick
time to got the correct delta -- animation works fine as a result.

Screen resizing code added -- basically we have a target width and
height and attempt to resize the canvas to the nearest width and height
divisible by said size. This isn't the best, however, as we should
probably, in addition to creating a base canvas width/height that is
ideal, also resize the total viewport to fill in the black borders.