Add sprite/object attach/detach, update rendering
Sprites and Objects may now attach/detach from each other. The code
implementing the object parenting is quite poor at the moment -- and the
same can be said about the hacked together physics loop.

To some extent, it seems that the separation of physics from rendering
is creating inefficiencies and bottlenecks.

All objects should now inherit from 'entity', as this implements the
base entity requirements.

Quite a bit needs to be refactored, but this should likely be approached
once the netcode is somewhat implemented.