createObject passes args, text sort of works
createObject(class, ...) now passes arguments to onConception(_, ...).

Text is semi-implemented. Need to create an optimized font sprite data
configuration -- something more efficient than specifying the frame
offsets for every character. Perhaps a C['f': 'm'] to assume that it is
a monospaced typeface and to assume width as repeating for x offsets.

Need to rewrite the engine to some extend -- likely fold the renderer
back into the main engine, as I am not going to implement a non-canvas
backend. This would allow Objects to call createSprite and similar
without having to prepend renderer.createSprite (thereby decreasing the
amount of characters used).