Add strange sprite animation editor
Basically an interface to editing the sprite.json file manually. It has
an animation preview and frame coordinates can be selected by using
the mouse on an HTML5 canvas. It also implements a Tree structure of
dubious quality.

There are many inefficiencies, BUT WHO CARES.

Oh, it also loads by dropping the .json and .png into the left and right
panels, respectively. Modified files can be saved via the 'Save' button.