Fix branch moveUp/moveDown, loading data, add canvas shiv
moveUp and moveDown now properly moves the branch's element and also
properly adjusts its position in its parent (if parent is also a
branch).

Sprite data was being saved with an {x,y,w,h,t} object per frame instead
of an array of x,y,w,h,t. Loading now accounts for sprite data in the
incorrect unoptimized format and saves frames as an array.

Added imageSmoothingEnabled shiv which likely doesn't work, as afaik
JavaScript properties are not by reference if they could be managed by
value (basically if they are a simple type, such as a boolean or number,
instead of an Object).