Add unfinished sprite packer and common.css file
packer.html implements a rudimentary sprite packing tool. At the moment
it takes in pairs of images and json files (spritesheets and their
accompanying sprite data) and attempts to make an optimized spritesheet
with all frames placed within it.

At the moment it does not provide modified data files, but the data is
there to do so. Lacking at the moment is processing for points, but that
shouldn't be too complicated to implement.

It will be expanded in the future to crop sprites to their smallest size
first -- this will require setting a "sprite origin" property so that
new frames will always be positioned at the appropriate location
relative to the previous frame.

common.css provides canvas settings and flexbox classes.